[Test] selected test for spec decode
