+++ tests/e2e/pull_request/light/coverage.md
@@ -75,6 +75,19 @@ The coverage of e2e light is as follows:
 | one_card/lora/test_qwen35_densemodel_lora.py | test_qwen35_text_lora | - | 1 |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  | ✅ | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |
 | one_card/lora/test_qwen3_multi_loras.py | test_multi_loras_with_tp_sync | - | 1 |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  | ✅ | ✅ | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |
 | one_card/lora/test_qwen3_reranker_lora.py | test_reranker_models_lora | - | 1 |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |
+| four_card/spec_decode/test_mtp_qwen3_next.py | test_qwen3_next_mtp_acceptance_tp4 | Qwen/Qwen3-Next-80B-A3B-Instruct | 4 |  |  | ✅ |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  |  |  | ✅ | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |
+| one_card/spec_decode/test_dflash.py | test_dflash_acceptance | Qwen/Qwen3-8B<br>z-lab/Qwen3-8B-DFlash-b16 | 1 |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |
+| one_card/spec_decode/test_draft_parallel.py | test_parallel_drafting_acceptance | LLM-Research/Meta-Llama-3.1-8B-Instruct<br>amd/PARD-Llama-3.2-1B | 1 |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |
+| one_card/spec_decode/test_eagle.py | test_qwen3_vl_eagle | Qwen/Qwen3-VL-8B-Instruct | 1 |  | ✅ |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  |  | ✅ |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |
+| one_card/spec_decode/test_eagle.py | test_qwen_eagle3_acceptance | Qwen/Qwen3-8B<br>RedHatAI/Qwen3-8B-speculator.eagle3 | 1 |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  |  |  | ✅ |  | ✅ |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |
+| one_card/spec_decode/test_extract_hidden_states.py | test_extract_hidden_states_eager_mode | Qwen/Qwen3-8B | 1 |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |
+| one_card/spec_decode/test_extract_hidden_states.py | test_extract_hidden_states_aclgraph_mode | Qwen/Qwen3-8B | 1 |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |
+| one_card/spec_decode/test_mtp_eagle_correctness.py | test_deepseek_mtp | wemaster/deepseek_mtp_main_random_bf16 | 1 |  |  | ✅ |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  |  |  | ✅ | ✅ |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |
+| one_card/spec_decode/test_ngram.py | test_ngram | LLM-Research/Meta-Llama-3.1-8B-Instruct | 1 |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |
+| one_card/spec_decode/test_ngram_npu.py | test_ngram_npu_async_acceptance | LLM-Research/Meta-Llama-3.1-8B-Instruct | 1 |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |
+| one_card/spec_decode/test_suffix.py | test_suffix_acceptance | LLM-Research/Meta-Llama-3.1-8B-Instruct | 1 |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |
+| two_card/spec_decode/test_spec_decode.py | test_eagle3_sp_acceptance | Qwen/Qwen3-8B<br>RedHatAI/Qwen3-8B-speculator.eagle3 | 2 |  | ✅ |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  |  |  | ✅ |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  | ✅ | ✅ |  |  |  |  |
+| two_card/spec_decode/test_spec_decode.py | test_p_eagle_acceptance | Qwen/Qwen3-Coder-30B-A3B-Instruct<br>amazon/Qwen3-Coder-30B-A3B-Instruct-P-EAGLE | 2 |  |  | ✅ |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  |  |  | ✅ |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |
 | one_card/test_batch_invariant.py | test_v1_generation_is_deterministic_across_batch_sizes_with_needle | Qwen/Qwen3-0.6B | 1 |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  | ✅ |  |
 | one_card/test_batch_invariant.py | test_logprobs_bitwise_batch_invariance_bs1_vs_bsN | Qwen/Qwen3-0.6B | 1 |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  | ✅ | ✅ |  |
 | one_card/test_batch_invariant.py | test_simple_generation | Qwen/Qwen3-0.6B | 1 |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  | ✅ |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |  |
