+++ tests/e2e/pull_request/light/four_card/spec_decode/test_mtp_qwen3_next.py
