+++ tests/e2e/pull_request/light/one_card/spec_decode/test_ngram.py
@@ -1,0 +1,26 @@
+from __future__ import annotations
+
+from typing import Any
+
+from vllm import SamplingParams
+
+from tests.e2e.conftest import VllmRunner
+
+
+def test_ngram(
+    test_prompts: list[list[dict[str, Any]]],
+    sampling_config: SamplingParams,
+    model_name: str,
+):
+    with VllmRunner(
+        model_name,
+        speculative_config={
+            "method": "ngram",
+            "prompt_lookup_max": 5,
+            "prompt_lookup_min": 3,
+            "num_speculative_tokens": 3,
+        },
+        max_model_len=1024,
+        cudagraph_capture_sizes=[1, 2, 4, 8],
+    ) as runner:
+        runner.model.chat(test_prompts, sampling_config)
